refactor build tests
