+++ dev/bots/test.dart
@@ -65,6 +65,7 @@ import 'package:process/process.dart';
 
 import 'run_command.dart';
 import 'suite_runners/run_add_to_app_life_cycle_tests.dart';
+import 'suite_runners/run_build_tests.dart';
 import 'suite_runners/run_flutter_packages_tests.dart';
 import 'suite_runners/run_realm_checker_tests.dart';
 import 'suite_runners/run_skp_generator_tests.dart';
@@ -231,7 +232,7 @@ Future<void> main(List<String> args) async {
     }
     await selectShard(<String, ShardRunner>{
       'add_to_app_life_cycle_tests': () => addToAppLifeCycleRunner(flutterRoot),
-      'build_tests': _runBuildTests,
+      'build_tests': () => buildTestsRunner(flutterRoot),
       'framework_coverage': _runFrameworkCoverage,
       'framework_tests': _runFrameworkTests,
       'tool_tests': _runToolTests,
@@ -504,296 +505,6 @@ Future<void> _runToolTests() async {
     'general': _runGeneralToolTests,
     'commands': _runCommandsToolTests,
   });
-}
-
-Future<void> runForbiddenFromReleaseTests() async {
-  // Build a release APK to get the snapshot json.
-  final Directory tempDirectory = Directory.systemTemp.createTempSync('flutter_forbidden_imports.');
-  final List<String> command = <String>[
-    'build',
-    'apk',
-    '--target-platform',
-    'android-arm64',
-    '--release',
-    '--analyze-size',
-    '--code-size-directory',
-    tempDirectory.path,
-    '-v',
-  ];
-
-  await runCommand(
-    flutter,
-    command,
-    workingDirectory: path.join(flutterRoot, 'examples', 'hello_world'),
-  );
-
-  // First, a smoke test.
-  final List<String> smokeTestArgs = <String>[
-    path.join(flutterRoot, 'dev', 'forbidden_from_release_tests', 'bin', 'main.dart'),
-    '--snapshot', path.join(tempDirectory.path, 'snapshot.arm64-v8a.json'),
-    '--package-config', path.join(flutterRoot, 'examples', 'hello_world', '.dart_tool', 'package_config.json'),
-    '--forbidden-type', 'package:flutter/src/widgets/framework.dart::Widget',
-  ];
-  await runCommand(
-    dart,
-    smokeTestArgs,
-    workingDirectory: flutterRoot,
-    expectNonZeroExit: true,
-  );
-
-  // Actual test.
-  final List<String> args = <String>[
-    path.join(flutterRoot, 'dev', 'forbidden_from_release_tests', 'bin', 'main.dart'),
-    '--snapshot', path.join(tempDirectory.path, 'snapshot.arm64-v8a.json'),
-    '--package-config', path.join(flutterRoot, 'examples', 'hello_world', '.dart_tool', 'package_config.json'),
-    '--forbidden-type', 'package:flutter/src/widgets/widget_inspector.dart::WidgetInspectorService',
-    '--forbidden-type', 'package:flutter/src/widgets/framework.dart::DebugCreator',
-    '--forbidden-type', 'package:flutter/src/foundation/print.dart::debugPrint',
-  ];
-  await runCommand(
-    dart,
-    args,
-    workingDirectory: flutterRoot,
-  );
-}
-
-/// Verifies that APK, and IPA (if on macOS), and native desktop builds the
-/// examples apps without crashing. It does not actually launch the apps. That
-/// happens later in the devicelab. This is just a smoke-test. In particular,
-/// this will verify we can build when there are spaces in the path name for the
-/// Flutter SDK and target app.
-///
-/// Also does some checking about types included in hello_world.
-Future<void> _runBuildTests() async {
-  final List<Directory> exampleDirectories = Directory(path.join(flutterRoot, 'examples')).listSync()
-    // API example builds will be tested in a separate shard.
-    .where((FileSystemEntity entity) => entity is Directory && path.basename(entity.path) != 'api').cast<Directory>().toList()
-    ..add(Directory(path.join(flutterRoot, 'packages', 'integration_test', 'example')))
-    ..add(Directory(path.join(flutterRoot, 'dev', 'integration_tests', 'android_semantics_testing')))
-    ..add(Directory(path.join(flutterRoot, 'dev', 'integration_tests', 'android_views')))
-    ..add(Directory(path.join(flutterRoot, 'dev', 'integration_tests', 'channels')))
-    ..add(Directory(path.join(flutterRoot, 'dev', 'integration_tests', 'hybrid_android_views')))
-    ..add(Directory(path.join(flutterRoot, 'dev', 'integration_tests', 'flutter_gallery')))
-    ..add(Directory(path.join(flutterRoot, 'dev', 'integration_tests', 'ios_platform_view_tests')))
-    ..add(Directory(path.join(flutterRoot, 'dev', 'integration_tests', 'ios_app_with_extensions')))
-    ..add(Directory(path.join(flutterRoot, 'dev', 'integration_tests', 'non_nullable')))
-    ..add(Directory(path.join(flutterRoot, 'dev', 'integration_tests', 'platform_interaction')))
-    ..add(Directory(path.join(flutterRoot, 'dev', 'integration_tests', 'spell_check')))
-    ..add(Directory(path.join(flutterRoot, 'dev', 'integration_tests', 'ui')));
-
-  // The tests are randomly distributed into subshards so as to get a uniform
-  // distribution of costs, but the seed is fixed so that issues are reproducible.
-  final List<ShardRunner> tests = <ShardRunner>[
-    for (final Directory exampleDirectory in exampleDirectories)
-      () => _runExampleProjectBuildTests(exampleDirectory),
-    ...<ShardRunner>[
-      // Web compilation tests.
-      () => _flutterBuildDart2js(
-            path.join('dev', 'integration_tests', 'web'),
-            path.join('lib', 'main.dart'),
-          ),
-      // Should not fail to compile with dart:io.
-      () => _flutterBuildDart2js(
-            path.join('dev', 'integration_tests', 'web_compile_tests'),
-            path.join('lib', 'dart_io_import.dart'),
-          ),
-      // Should be able to compile with a call to:
-      // BackgroundIsolateBinaryMessenger.ensureInitialized.
-      () => _flutterBuildDart2js(
-            path.join('dev', 'integration_tests', 'web_compile_tests'),
-            path.join('lib', 'background_isolate_binary_messenger.dart'),
-          ),
-    ],
-    runForbiddenFromReleaseTests,
-  ]..shuffle(math.Random(0));
-
-  await runShardRunnerIndexOfTotalSubshard(tests);
-}
-
-Future<void> _runExampleProjectBuildTests(Directory exampleDirectory, [File? mainFile]) async {
-  // Only verify caching with flutter gallery.
-  final bool verifyCaching = exampleDirectory.path.contains('flutter_gallery');
-  final String examplePath = path.relative(exampleDirectory.path, from: Directory.current.path);
-  final List<String> additionalArgs = <String>[
-    if (mainFile != null) path.relative(mainFile.path, from: exampleDirectory.absolute.path),
-  ];
-  if (Directory(path.join(examplePath, 'android')).existsSync()) {
-    await _flutterBuildApk(examplePath, release: false, additionalArgs: additionalArgs, verifyCaching: verifyCaching);
-    await _flutterBuildApk(examplePath, release: true, additionalArgs: additionalArgs, verifyCaching: verifyCaching);
-  } else {
-    print('Example project ${path.basename(examplePath)} has no android directory, skipping apk');
-  }
-  if (Platform.isMacOS) {
-    if (Directory(path.join(examplePath, 'ios')).existsSync()) {
-      await _flutterBuildIpa(examplePath, release: false, additionalArgs: additionalArgs, verifyCaching: verifyCaching);
-      await _flutterBuildIpa(examplePath, release: true, additionalArgs: additionalArgs, verifyCaching: verifyCaching);
-    } else {
-      print('Example project ${path.basename(examplePath)} has no ios directory, skipping ipa');
-    }
-  }
-  if (Platform.isLinux) {
-    if (Directory(path.join(examplePath, 'linux')).existsSync()) {
-      await _flutterBuildLinux(examplePath, release: false, additionalArgs: additionalArgs, verifyCaching: verifyCaching);
-      await _flutterBuildLinux(examplePath, release: true, additionalArgs: additionalArgs, verifyCaching: verifyCaching);
-    } else {
-      print('Example project ${path.basename(examplePath)} has no linux directory, skipping Linux');
-    }
-  }
-  if (Platform.isMacOS) {
-    if (Directory(path.join(examplePath, 'macos')).existsSync()) {
-      await _flutterBuildMacOS(examplePath, release: false, additionalArgs: additionalArgs, verifyCaching: verifyCaching);
-      await _flutterBuildMacOS(examplePath, release: true, additionalArgs: additionalArgs, verifyCaching: verifyCaching);
-    } else {
-      print('Example project ${path.basename(examplePath)} has no macos directory, skipping macOS');
-    }
-  }
-  if (Platform.isWindows) {
-    if (Directory(path.join(examplePath, 'windows')).existsSync()) {
-      await _flutterBuildWin32(examplePath, release: false, additionalArgs: additionalArgs, verifyCaching: verifyCaching);
-      await _flutterBuildWin32(examplePath, release: true, additionalArgs: additionalArgs, verifyCaching: verifyCaching);
-    } else {
-      print('Example project ${path.basename(examplePath)} has no windows directory, skipping Win32');
-    }
-  }
-}
-
-Future<void> _flutterBuildApk(String relativePathToApplication, {
-  required bool release,
-  bool verifyCaching = false,
-  List<String> additionalArgs = const <String>[],
-}) async {
-  printProgress('${green}Testing APK ${release ? 'release' : 'debug'} build$reset for $cyan$relativePathToApplication$reset...');
-  await _flutterBuild(relativePathToApplication, 'APK', 'apk',
-    release: release,
-    verifyCaching: verifyCaching,
-    additionalArgs: additionalArgs
-  );
-}
-
-Future<void> _flutterBuildIpa(String relativePathToApplication, {
-  required bool release,
-  List<String> additionalArgs = const <String>[],
-  bool verifyCaching = false,
-}) async {
-  assert(Platform.isMacOS);
-  printProgress('${green}Testing IPA ${release ? 'release' : 'debug'} build$reset for $cyan$relativePathToApplication$reset...');
-  await _flutterBuild(relativePathToApplication, 'IPA', 'ios',
-    release: release,
-    verifyCaching: verifyCaching,
-    additionalArgs: <String>[...additionalArgs, '--no-codesign'],
-  );
-}
-
-Future<void> _flutterBuildLinux(String relativePathToApplication, {
-  required bool release,
-  bool verifyCaching = false,
-  List<String> additionalArgs = const <String>[],
-}) async {
-  assert(Platform.isLinux);
-  await runCommand(flutter, <String>['config', '--enable-linux-desktop']);
-  printProgress('${green}Testing Linux ${release ? 'release' : 'debug'} build$reset for $cyan$relativePathToApplication$reset...');
-  await _flutterBuild(relativePathToApplication, 'Linux', 'linux',
-    release: release,
-    verifyCaching: verifyCaching,
-    additionalArgs: additionalArgs
-  );
-}
-
-Future<void> _flutterBuildMacOS(String relativePathToApplication, {
-  required bool release,
-  bool verifyCaching = false,
-  List<String> additionalArgs = const <String>[],
-}) async {
-  assert(Platform.isMacOS);
-  await runCommand(flutter, <String>['config', '--enable-macos-desktop']);
-  printProgress('${green}Testing macOS ${release ? 'release' : 'debug'} build$reset for $cyan$relativePathToApplication$reset...');
-  await _flutterBuild(relativePathToApplication, 'macOS', 'macos',
-    release: release,
-    verifyCaching: verifyCaching,
-    additionalArgs: additionalArgs
-  );
-}
-
-Future<void> _flutterBuildWin32(String relativePathToApplication, {
-  required bool release,
-  bool verifyCaching = false,
-  List<String> additionalArgs = const <String>[],
-}) async {
-  assert(Platform.isWindows);
-  printProgress('${green}Testing ${release ? 'release' : 'debug'} Windows build$reset for $cyan$relativePathToApplication$reset...');
-  await _flutterBuild(relativePathToApplication, 'Windows', 'windows',
-    release: release,
-    verifyCaching: verifyCaching,
-    additionalArgs: additionalArgs
-  );
-}
-
-Future<void> _flutterBuild(
-  String relativePathToApplication,
-  String platformLabel,
-  String platformBuildName, {
-  required bool release,
-  bool verifyCaching = false,
-  List<String> additionalArgs = const <String>[],
-}) async {
-  await runCommand(flutter,
-    <String>[
-      'build',
-      platformBuildName,
-      ...additionalArgs,
-      if (release)
-        '--release'
-      else
-        '--debug',
-      '-v',
-    ],
-    workingDirectory: path.join(flutterRoot, relativePathToApplication),
-  );
-
-  if (verifyCaching) {
-    printProgress('${green}Testing $platformLabel cache$reset for $cyan$relativePathToApplication$reset...');
-    await runCommand(flutter,
-      <String>[
-        'build',
-        platformBuildName,
-        '--performance-measurement-file=perf.json',
-        ...additionalArgs,
-        if (release)
-          '--release'
-        else
-          '--debug',
-        '-v',
-      ],
-      workingDirectory: path.join(flutterRoot, relativePathToApplication),
-    );
-    final File file = File(path.join(flutterRoot, relativePathToApplication, 'perf.json'));
-    if (!_allTargetsCached(file)) {
-      foundError(<String>[
-        '${red}Not all build targets cached after second run.$reset',
-        'The target performance data was: ${file.readAsStringSync().replaceAll('},', '},\n')}',
-      ]);
-    }
-  }
-}
-
-bool _allTargetsCached(File performanceFile) {
-  final Map<String, Object?> data = json.decode(performanceFile.readAsStringSync())
-    as Map<String, Object?>;
-  final List<Map<String, Object?>> targets = (data['targets']! as List<Object?>)
-    .cast<Map<String, Object?>>();
-  return targets.every((Map<String, Object?> element) => element['skipped'] == true);
-}
-
-Future<void> _flutterBuildDart2js(String relativePathToApplication, String target, { bool expectNonZeroExit = false }) async {
-  printProgress('${green}Testing Dart2JS build$reset for $cyan$relativePathToApplication$reset...');
-  await runCommand(flutter,
-    <String>['build', 'web', '-v', '--target=$target'],
-    workingDirectory: path.join(flutterRoot, relativePathToApplication),
-    expectNonZeroExit: expectNonZeroExit,
-    environment: <String, String>{
-      'FLUTTER_WEB': 'true',
-    },
-  );
 }
 
 Future<void> _runFrameworkTests() async {
